+++ app/Http/Controllers/TaskController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
 use App\Models\TaskModel;
-use App\Models\SubTaskModel;
+use App\Http\Controllers\AccountController;
 
 class TaskController extends Controller
 {
@@ -32,7 +32,7 @@ class TaskController extends Controller
         $tasks = DB::table('tasks')
             ->leftJoin('employees', 'tasks.employee_id', '=', 'employees.employee_id')
             ->select('tasks.*', 'employees.last_name', 'employees.first_name')
-            ->where('project_id', $request->id)
+            ->where('project_location_id', $request->id)
             ->get();
         return response()->json(['success' => true,'tasks' => $tasks]);
     }
@@ -53,6 +53,7 @@ class TaskController extends Controller
     public function create(Request $request)
     {
         try {
+            
             $validatedData = $request->validate([
                 'id' => 'required|string',
                 'taskname' => 'required|string|max:255',
@@ -68,14 +69,44 @@ class TaskController extends Controller
             $empl_allow = DB::table('employees')->select('employees.employee_id')
                 ->join('team_details', 'employees.employee_id', '=', 'team_details.employee_id')
                 ->join('project_teams', 'team_details.team_id', '=', 'project_teams.team_id')
-                ->where('project_teams.project_id', $validatedData['id'])->get()->toArray();
-            if(!in_array($validatedData['employee_id'], $empl_allow)){
+                ->join('project_locations', 'project_teams.project_id', '=', 'project_locations.project_id')
+                ->where('project_locations.project_location_id', $validatedData['id'])
+                ->get()->toArray();
+            $empl_allow_ids = array_map(function($empl) {
+                    return $empl->employee_id;
+            }, $empl_allow);
+            if(!in_array($validatedData['employee_id'], $empl_allow_ids)){
                 return response()->json(['success' => false,'message' => 'Employee is not in this project']);
+            }
+            $leaders = DB::table('employees')->select('employees.employee_id')
+            ->join('team_details', 'employees.employee_id', '=', 'team_details.employee_id')
+            ->join('project_teams', 'team_details.team_id', '=', 'project_teams.team_id')
+            ->join('project_locations', 'project_teams.project_id', '=', 'project_locations.project_id')
+            ->where('project_locations.project_location_id', $validatedData['id'])
+            ->where('team_details.team_permission', 1)
+            ->get();
+            $data = \Illuminate\Support\Facades\DB::table('accounts')
+                ->join('employees', 'accounts.employee_id', '=', 'employees.employee_id')
+                ->join('contacts', 'employees.contact_id', '=', 'contacts.contact_id')
+                ->join('job_details', 'job_details.employee_id', '=', 'employees.employee_id')
+
+                ->where(
+                    'accounts.account_id',
+                    \Illuminate\Support\Facades\Request::session()->get(\App\StaticString::ACCOUNT_ID),
+                )
+                ->first();
+            $empl_id = (int) $data->employee_id;
+            $leader_arr = [];
+            foreach ($leaders as $l) {
+                array_push($leader_arr, $l->employee_id);
+            }
+            if (!in_array($empl_id, $leader_arr) && !in_array(AccountController::permissionStr(), ['super', 'admin', 'project_manager'])){
+                return response()->json(['success' => false,'message' => 'You do not have permission to add task']);
             }
 
             $task = TaskModel::create([
                 'task_name' => $validatedData['taskname'],
-                'project_id' => $validatedData['id'],
+                'project_location_id' => $validatedData['id'],
                 'request' => $validatedData['request'],
                 'employee_id' => $validatedData['employee_id'],
                 'start_date' => $validatedData['s_date'],
@@ -91,7 +122,7 @@ class TaskController extends Controller
                     TaskModel::create([
                         'task_name' => $value,
                         'employee_id' => $validatedData['employee_id'],
-                        'project_id' => $validatedData['id'],
+                        'project_location_id' => $validatedData['id'],
                         'start_date' => $validatedData['s_date'],
                         'end_date' => $validatedData['s_date'],
                         'parent_id' => $task->task_id,
@@ -115,7 +146,7 @@ class TaskController extends Controller
             $tasks = DB::table('tasks')
                 ->leftJoin('employees', 'tasks.employee_id', '=', 'employees.employee_id')
                 ->select('tasks.*', 'employees.last_name', 'employees.first_name')
-                ->where('project_id', $validatedData['id'])
+                ->where('project_location_id', $validatedData['id'])
                 ->get();
             return response()->json(['success' => true,'message' => 'Add new task success','tasks' => $tasks]);
         } catch (\Exception $e) {
@@ -135,18 +166,38 @@ class TaskController extends Controller
                 's_date' => 'required|date',
                 'e_date' => 'required|date|after_or_equal:s_date',
             ]);
+            $data = \Illuminate\Support\Facades\DB::table('accounts')
+                ->join('employees', 'accounts.employee_id', '=', 'employees.employee_id')
+                ->join('contacts', 'employees.contact_id', '=', 'contacts.contact_id')
+                ->join('job_details', 'job_details.employee_id', '=', 'employees.employee_id')
+                ->where(
+                    'accounts.account_id',
+                    \Illuminate\Support\Facades\Request::session()->get(\App\StaticString::ACCOUNT_ID),
+                )
+                ->first();
+            $empl_id = (int) $data->employee_id;
+
+            $empl_allow = DB::table('employees')->select('employees.employee_id')
+                ->join('team_details', 'employees.employee_id', '=', 'team_details.employee_id')
+                ->join('project_teams', 'team_details.team_id', '=', 'project_teams.team_id')
+                ->join('project_locations', 'project_teams.project_id', '=', 'project_locations.project_id')
+                ->where('project_locations.project_location_id', $validatedData['id'])
+                ->get()->toArray();
+            $empl_allow_ids = array_map(function($empl) {
+                    return $empl->employee_id;
+            }, $empl_allow);
+            if (!in_array($empl_id, $empl_allow_ids)){
+                return response()->json(['success' => false,'message' => 'You do not have permission to update task']);
+            }
+
+            if(!in_array($validatedData['employee_id'], $empl_allow_ids)){
+                return response()->json(['success' => false,'message' => 'Employee is not in this project']);
+            }
+
             $progress = 0;
             if($request->has('allmarkdone')){
                 $progress = 1;
             }
-            $empl_allow = DB::table('employees')->select('employees.employee_id')
-            ->join('team_details', 'employees.employee_id', '=', 'team_details.employee_id')
-            ->join('project_teams', 'team_details.team_id', '=', 'project_teams.team_id')
-            ->where('project_teams.project_id', $validatedData['id'])->get()->toArray();
-            if(!in_array($validatedData['employee_id'], $empl_allow)){
-                return response()->json(['success' => false,'message' => 'Employee is not in this project']);
-            }
-
             TaskModel::where('task_id', $validatedData['task_id'])->update([
                 'task_name' => $validatedData['taskname'],
                 'request' => $validatedData['request'],
@@ -169,7 +220,7 @@ class TaskController extends Controller
                         $sub=TaskModel::create([
                             'task_name' => $value,
                             'parent_id' => $validatedData['task_id'],
-                            'project_id' => $validatedData['id'],
+                            'project_location_id' => $validatedData['id'],
                             'progress' => $progress,
                         ]);
                         array_push($subtask_id, $sub->task_id);
@@ -197,11 +248,24 @@ class TaskController extends Controller
                     ]);
                 }
             }
+            $par_id=TaskModel::where('task_id', $validatedData['task_id'])->pluck('parent_id')->first();
+            if(!empty($par_id)){
+                $subtasks = DB::table('tasks')->where('parent_id', $par_id)->get();
+                if(count($subtasks)){
+                    $progress = 0;
+                    foreach ($subtasks as $subtask) {
+                        $progress += $subtask->progress;
+                    }
+                    TaskModel::where('task_id', $par_id)->update([
+                        'progress' => $progress / count($subtasks),
+                    ]);
+                }
+            }
 
             $tasks = DB::table('tasks')
                 ->leftJoin('employees', 'tasks.employee_id', '=', 'employees.employee_id')
                 ->select('tasks.*', 'employees.last_name', 'employees.first_name')
-                ->where('project_id', $validatedData['id'])
+                ->where('project_location_id', $validatedData['id'])
                 ->get();
             return response()->json(['success' => true,'message' => 'Update task success','tasks' => $tasks]);
         } catch (\Exception $e) {
@@ -212,6 +276,32 @@ class TaskController extends Controller
     public function delete(Request $request)
     {
         try {
+            $location_id = TaskModel::where('task_id', $request->id)->pluck('project_location_id')->first();
+            $leaders = DB::table('employees')->select('employees.employee_id')
+            ->join('team_details', 'employees.employee_id', '=', 'team_details.employee_id')
+            ->join('project_teams', 'team_details.team_id', '=', 'project_teams.team_id')
+            ->join('project_locations', 'project_teams.project_id', '=', 'project_locations.project_id')
+            ->where('project_locations.project_location_id', $location_id)
+            ->where('team_details.team_permission', 1)
+            ->get();
+            $data = \Illuminate\Support\Facades\DB::table('accounts')
+                ->join('employees', 'accounts.employee_id', '=', 'employees.employee_id')
+                ->join('contacts', 'employees.contact_id', '=', 'contacts.contact_id')
+                ->join('job_details', 'job_details.employee_id', '=', 'employees.employee_id')
+
+                ->where(
+                    'accounts.account_id',
+                    \Illuminate\Support\Facades\Request::session()->get(\App\StaticString::ACCOUNT_ID),
+                )
+                ->first();
+            $empl_id = (int) $data->employee_id;
+            $leader_arr = [];
+            foreach ($leaders as $l) {
+                array_push($leader_arr, $l->employee_id);
+            }
+            if (!in_array($empl_id, $leader_arr) && !in_array(AccountController::permissionStr(), ['super', 'admin', 'project_manager'])){
+                return response()->json(['success' => false,'message' => 'You do not have permission to delete task']);
+            }
             TaskModel::find($request->id)->delete();
             TaskModel::where('parent_id', $request->id)->delete();
